Second pass at moving to ViewPager2
