+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/tutorial/CardReaderTutorialDialogFragment.kt
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.Window
 import androidx.fragment.app.DialogFragment
+import androidx.viewpager2.widget.ViewPager2
 import com.woocommerce.android.R
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.databinding.DialogCardReaderTutorialBinding
@@ -33,13 +34,14 @@ class CardReaderTutorialDialogFragment : DialogFragment(R.layout.dialog_card_rea
 
         val binding = DialogCardReaderTutorialBinding.bind(view)
 
-        binding.viewPager.initViewPager(this)
-        /*binding.viewPager.addOnPageChangeListener(object : ViewPager.SimpleOnPageChangeListener() {
+        binding.viewPager.adapter = CardReaderTutorialAdapter(this)
+        binding.viewPager.registerOnPageChangeCallback(object : ViewPager2.OnPageChangeCallback() {
             override fun onPageSelected(position: Int) {
-                val lastPosition = CardReaderTutorialViewPager.NUM_PAGES - 1
+                val lastPosition = CardReaderTutorialAdapter.NUM_PAGES - 1
                 binding.buttonSkip.setText(if (position == lastPosition) R.string.close else R.string.skip)
             }
-        })*/
+        })
+        binding.viewPagerIndicator.setupFromViewPager(binding.viewPager)
 
         binding.buttonSkip.setOnClickListener {
             navigateBackWithResult(KEY_READER_TUTORIAL_RESULT, true)
